% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{check}
\alias{check}
\title{Check password_hash_table}
\usage{
check(x)
}
\arguments{
\item{x}{}
}
\description{
Checks if x is in password_hash_table$alnum
and returns the encrypted value if it finds it, else it
returns x
}
