% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{encrypt}
\alias{encrypt}
\title{Encrypt}
\usage{
encrypt(...)
}
\arguments{
\item{...}{}
}
\description{
Takes in a character vector and returns an
encrypted string based on the password_hash_table
available in package
}
