% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encrypt.R
\name{encrypt_char}
\alias{encrypt_char}
\title{Encrypt a character vector using an RSA public/private key}
\usage{
encrypt_char(.data, public_key_path = "id_rsa.pub")
}
\arguments{
\item{.data}{A character vector.}

\item{public_key_path}{Character. A quoted path to an RSA public key created
using \code{\link{write_keys}}.}
}
\value{
A vector of ciphertexts.
}
\description{
Not usually called directly.
}
\examples{
\dontrun{
hospital_number = c("1010761111", "2010761212")
encrypt_char(hospital_number)
}
}
