% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptr_package.R
\docType{package}
\name{encryptr-package}
\alias{encryptr-package}
\title{encryptr: Encrypt and decrypt data frame or tibble columns using the strong RSA public/private keys.}
\description{
encryptr: Encrypt and decrypt data frame or tibble columns using the strong RSA public/private keys.
}
\section{\code{encryptr} key generation}{


\link{genkeys},
}

\section{\code{encryptr} encrypt/decrypt}{


\link{encrypt},
\link{decrypt}
}

