% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encryptr_package.R
\docType{data}
\name{gp}
\alias{gp}
\title{General Practitioner (family doctor) practices in Scotland 2018}
\format{
A data frame with 1212 rows and 12 variables
}
\usage{
data(gp)
}
\description{
From NHS Digital Organisational Services
Downloaded February 2019
}
\keyword{data}
