% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{load}
\alias{load}
\title{Load encrypted R objects}
\usage{
load(file, envir = parent.frame(), password = NULL)
}
\arguments{
\item{file}{the path to the file.}

\item{envir}{the environment to attach to.}

\item{password}{the password to use, if NULL, the function asks for a password.}
}
\description{
This function allows you to load encrypted R objects.
Loaded R objects are attached to the environment.
}
\details{
Uses AES 256-bit encryption (in CBC mode) with a SHA-256 hashed password.
}
\note{
Asking for a password relies on being in interactive() mode (RStudio, RGui).
Outside interactive() mode (rmarkdown, knitr), the package "getPass" is required.
}
\examples{
load("mydata.Rdata")

}
