% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{save}
\alias{save}
\title{Save R objects as an encrypted file.}
\usage{
save(..., list = character(), file = stop("'file' must be specified"),
  envir = parent.frame(), password = NULL)
}
\arguments{
\item{...}{the variables to be saved.}

\item{list}{an optional list of names that identity the R objects to store.}

\item{file}{the path to the file.}

\item{envir}{the environment the variables are token from.}

\item{password}{the password to use, if NULL, the package tries to ask for a password.}
}
\description{
This function allows you to save encrypted R objects.
R objects are taken from the environment.
}
\details{
Uses AES 256-bit encryption (in CBC mode) with a SHA-256 hashed password.
}
\note{
Asking for a password relies on being in interactive() mode (RStudio, RGui).
Outside interactive() mode (rmarkdown, knitr), the package "getPass" is required.
}
\examples{
df <- data.frame(rnorm(100))
save(df, file="mydata.Rdata")

}
